requirements.txt
openai
requests
pdfplumber
beautifulsoup4
googlesearch-python
sectxt
python-dotenv